selection: Separate replacement preparation from publication
Transformed line selections bind the original change to rewritten file
content before saving ownership in a batch.

I need to audit the selection decision separately from buffer lifetime and
batch publication. The preparation function currently carries those
responsibilities through hundreds of lines of temporary flags.

This commit separates selection policy, projection, rendering, parent
boundaries, and storage into their owning modules. A resolved decision
carries the content strategy into preparation, and callers import
publication from storage.
